+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.0] - unreleased
 
+### Added
+
+- Optional command line parameter `-ld` for specifying source library directory.
+
 ## [1.0.0] - 2024-12-12
 
 Initial public release.